+++ packages/lumx-react/src/components/date-picker/DatePicker.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { WrappedDatePicker } from '@lumx/react';
+import { DatePickerField } from '@lumx/react';
 
 import { decorators } from '@lumx/react/story-block';
 
@@ -9,5 +9,5 @@ export default { title: 'DatePicker', decorators };
 export const simpleDatePicker = ({ theme }) => {
     const [value, setValue] = React.useState();
 
-    return <WrappedDatePicker locale="fr" label="Start date" theme={theme} onChange={setValue} value={value} />;
+    return <DatePickerField locale="fr" label="Start date" theme={theme} onChange={setValue} value={value} />;
 };
